[Ent Server 10.2.0] [EP-2333] Repair bug introduced during last rework. Detect silent errors generating Java files by checking existence of re-generated service locator file. Add the SystemAdmin interface (was forgotten in the first place).
Former-commit-id: 02b2723b7ac9644dcdff301837cbc150731691a3

+++ Enterprise/Enterprise/server/buildtools/genservices/WebServiceClassesGenerator.class.php
@@ -425,7 +425,7 @@ class WW_BuildTools_GenServices_WebServiceClassesGenerator
 		$intfShortLow = strtolower( $this->intfDescriptor->getServiceNameShort() );
 
 		// Delete output folder recursively (if exists).
-		$outputFolder = $this->composeSdkDirForLanguage( 'java' )."/${$intfShortLow}";
+		$outputFolder = $this->composeSdkDirForLanguage( 'java' )."/{$intfShortLow}";
 		if( file_exists( $outputFolder ) ) {
 			require_once BASEDIR.'/server/utils/FolderUtils.class.php';
 			FolderUtils::cleanDirRecursive( $outputFolder, false );
@@ -458,7 +458,13 @@ class WW_BuildTools_GenServices_WebServiceClassesGenerator
 		$defaultEntryPoint = str_replace( LOCALURL_ROOT.INETROOT.'/', 'http://127.0.0.1/Enterprise/', $entryPoint );
 		list( $prefix, $nameSpace ) = explode( ':', $this->intfDescriptor->getNameSpace(), 2 ); // namespace has 'urn:' prefix
 		$javaFile = $outputFolder.'/'.$nameSpace.'ServiceLocator.java';
-		file_put_contents( $javaFile, str_replace( $entryPoint, $defaultEntryPoint, file_get_contents( $javaFile ) ) );
+		if( !file_exists( $javaFile ) ) {
+			$this->FatalErrors[] = 'Could not generate Java classes. The service locator file is missing in the output: '.
+				$javaFile.PHP_EOL.
+				'Please fix it and try again. You could manually retry as follows: '.$command.PHP_EOL;
+		} else {
+			file_put_contents( $javaFile, str_replace( $entryPoint, $defaultEntryPoint, file_get_contents( $javaFile ) ) );
+		}
 	}
 
 	/**
